pyproject.toml
[project]
name = "intelligent-evidence-platform"
version = "0.1.0"
description = "智能证据平台 - 法律债务纠纷领域的证据智能管理平台"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "asyncpg>=0.28.0",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "pydantic>=2.4.0",
    "pydantic-settings>=2.0.3",
    "pgvector>=0.2.3",
    "cos-python-sdk-v5>=1.9.25",
    "uvicorn>=0.23.2",
    "greenlet>=3.2.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.5.1",
    "ruff>=0.0.287",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 88
target-version = "py312"
select = ["E", "F", "B", "I", "C4"]
ignore = []
